refactor: Split search page into more components
This enables the decoupling of the state between components which means
that the input value can be re re-rendered independently of the fetch
request from the search method.  Without this decoupling, the input
value didn't update until the book results had been fetched, making
typing a string in the search box painfully slow.