refactor: Make a single call to setState in App
At this point, I'm sure if it's all that beneficial to utilise a temp
object that I can manipulate and then pass to setState once, rather than
calling setState multiple times for each shelf. This is probably a more
efficient way of doing it, but whether it makes a tangible difference,
it's hard to tell!