fix: Search page correctly displays books' shelves
There was a bug where if the search page was loaded directly rather than
navigating to it from the home page, the books' shelf state would not be
rendered correctly.  This resulted in all books in the search page
having the apparent shelf of "none" even when they were actually on a
shelf.

Changing the point at which the search page retrieved and stored the
books' identifiers from the props, meant that these ids are now
correctly stored, fixing the bug.