fix: Fix bug where shelves don't render correctly
Having the scoped object bookIds in the SearchBooks component was
causing issues.  After reading around a bit, it seems that trying to
combine scoped variables and component lifecycle hooks can get messy!
To alleviate these issues, I've moved the bookIds object, and the logic
for assigning values to bookIds, into the updateSearchResults function.

Another solution would be to store these bookIds in the state of
SearchBooks, but I feel that that tactic would be a bit heavy handed,
as bookIds doesn't really reflect the state of the component, it's more
of an abstraction of the books prop that is passed into the component.