fix: Ensure search results are correct
Because events from the input of the keyboard are triggered
asynchronously, state can be changed in a different order than that
which is entered.  This is especially problematic when results that are
returned should be blank.  Because the computation time of processing an
empty set of results is far shorter than when processing a full list of
20 book results, this can result in a "blank" result being triggered
after an event that returns results, but being completed BEFORE the
more intensive process.  The result is that empty search strings will
appear to show results, especially with rapid user input of the
keyboard.

For example:

* Entering "j" returns results for "JavaScript"
* Entering "jo" returns results for "Journey"
* Entering "journey" returns results for "Journey", but...
* Holding down the delete key deleting all letters of "journey" results
    in results for "JavaScript"

This is because the search result for "" (blank) finishing processing
before the results of "j" (JavaScript) can be fetched and processed.

To get around this, the searchQuery state in the SearchBar component has
been moved up to the SearchPage component (currentSearchQuery), and is
used as a final check before the results are updated in the UI.  If the
currentSearchQuery matches the query passed in to attemptResultUpdate,
then the query and the results fetched are still equivalent, and so the
UI can be updated.  If these two parameters are not equivalent, then a
new query has been entered by the user in the meantime, and the
processed results can be discarded.

This results in the results always displaying the correct books for the
current search query.